% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{bg_bind_debug_panel}
\alias{bg_bind_debug_panel}
\title{Lier le panneau de debug}
\usage{
bg_bind_debug_panel(output, journal_reactive)
}
\arguments{
\item{output}{Objet \code{output} de Shiny.}

\item{journal_reactive}{Réactif (\code{reactive()}/\code{reactiveVal()}) retournant un data.frame de jobs.}
}
\description{
Lie la sortie UI \code{output$bg_debug_panel} à un journal réactif.
}
