% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{bg_debug_panel}
\alias{bg_debug_panel}
\title{Debug panel (UI) for job journal}
\usage{
bg_debug_panel(journal_reactive)
}
\arguments{
\item{journal_reactive}{A reactive() returning the jobs data.frame}
}
\description{
Debug panel (UI) for job journal
}
