% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifier.R
\name{bg_make_notifier}
\alias{bg_make_notifier}
\title{Construire un notifier}
\usage{
bg_make_notifier(
  mode = c("none", "shiny", "message"),
  verbosity = c("silent", "info", "debug")
)
}
\arguments{
\item{mode}{"none" | "shiny" | "message" | function(type, text)}

\item{verbosity}{"silent" | "info" | "debug"}
}
\value{
une fonction notifier(type, text, level = "info")
}
\description{
Construire un notifier
}
