% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifier.R
\name{.bg_now_ms}
\alias{.bg_now_ms}
\title{Milliseconds since epoch (helper)}
\usage{
.bg_now_ms()
}
\value{
numeric milliseconds (double)
}
\description{
Milliseconds since epoch (helper)
}
\keyword{internal}
