% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timevis_log.R
\name{tv_widget}
\alias{tv_widget}
\title{Assembler le widget \pkg{timevis}}
\usage{
tv_widget(
  items,
  groups,
  add_css = TRUE,
  add_legend = TRUE,
  options = list(stack = TRUE, showCurrentTime = TRUE, zoomKey = "altKey")
)
}
\arguments{
\item{items}{Data.frame \code{tv$items} (depuis \code{tv_prepare()}).}

\item{groups}{Data.frame \code{tv$groups} (depuis \code{tv_prepare()}).}

\item{add_css}{Injecter le CSS des drapeaux/traits.}

\item{add_legend}{Ajouter la légende.}

\item{options}{Options passées à \code{timevis::timevis()}.}
}
\value{
Un \code{htmlwidgets} prêt à afficher.
}
\description{
Assembler le widget \pkg{timevis}
}
