% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{bg_bind_debug_panel}
\alias{bg_bind_debug_panel}
\title{Server binder for debug panel}
\usage{
bg_bind_debug_panel(output, journal_reactive)
}
\arguments{
\item{output}{Shiny output list (server-side)}

\item{journal_reactive}{A reactive() returning the jobs data.frame}
}
\description{
Server binder for debug panel
}
